move tests to a new file tag-from-scope
